fix: Move tests to Repository
