+++ phpunit.xml
@@ -68,9 +68,6 @@
     <testsuite name="eZ\Publish\API\Repository\Values\Content">
       <directory>eZ/Publish/API/Repository/Tests/Values/Content</directory>
     </testsuite>
-    <testsuite name="eZ\Publish\Core\Repository\Strategy\ContentThumbnail">
-      <directory>eZ/Publish/Core/Repository/Strategy/ContentThumbnail/Tests</directory>
-    </testsuite>
   </testsuites>
   <filter>
     <whitelist>
